requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
python-dotenv
werkzeug==0.16
requests
